[#84] feat: 직원 사용량 조회 기능 추가

+++ backend/src/main/java/com/e201/api/controller/company/request/employee/EmployeeCreateRequest.java
@@ -25,10 +25,10 @@ public class EmployeeCreateRequest {
 	private String name;
 
 	@NotBlank
-	private Integer supportAmount;
+	private Long supportAmount;
 
 	@Builder
-	public EmployeeCreateRequest(String code, String password, String name, Integer supportAmount) {
+	public EmployeeCreateRequest(String code, String password, String name, Long supportAmount) {
 		this.code = code;
 		this.password = password;
 		this.name = name;
